requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pyaudio>=0.2.13
openai>=1.55.0
pvporcupine>=3.0.0
numpy>=1.24.0,<1.25.0
pydantic>=2.5.2
pydantic-settings>=2.1.0
aiofiles>=23.2.1
RPi.GPIO>=0.7.1
python-dotenv>=1.0.0
sounddevice>=0.4.6
psutil>=5.9.0 